feat(update): add update state display for all cards/chips
