+++ src/cards/chips-card/chips/light-chip.ts
@@ -1,14 +1,9 @@
-import {
-    ActionHandlerEvent,
-    computeStateDisplay,
-    handleAction,
-    hasAction,
-    HomeAssistant,
-} from "custom-card-helpers";
+import { ActionHandlerEvent, handleAction, hasAction, HomeAssistant } from "custom-card-helpers";
 import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 import { classMap } from "lit/directives/class-map.js";
 import { styleMap } from "lit/directives/style-map.js";
+import { computeStateDisplay } from "../../../utils/compute-state-display";
 import { actionHandler } from "../../../utils/directives/action-handler-directive";
 import { isActive } from "../../../utils/entity";
 import { stateIcon } from "../../../utils/icons/state-icon";
