+++ src/cards/chips-card/chips/weather-chip.ts
@@ -1,6 +1,5 @@
 import {
     ActionHandlerEvent,
-    computeStateDisplay,
     formatNumber,
     handleAction,
     hasAction,
@@ -8,6 +7,7 @@ import {
 } from "custom-card-helpers";
 import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
+import { computeStateDisplay } from "../../../utils/compute-state-display";
 import { actionHandler } from "../../../utils/directives/action-handler-directive";
 import {
     computeChipComponentName,
