+++ src/cards/light-card/light-card.ts
@@ -1,6 +1,5 @@
 import {
     ActionHandlerEvent,
-    computeStateDisplay,
     handleAction,
     hasAction,
     HomeAssistant,
@@ -18,6 +17,7 @@ import "../../shared/shape-icon";
 import "../../shared/state-info";
 import "../../shared/state-item";
 import { cardStyle } from "../../utils/card-styles";
+import { computeStateDisplay } from "../../utils/compute-state-display";
 import { registerCustomCard } from "../../utils/custom-cards";
 import { actionHandler } from "../../utils/directives/action-handler-directive";
 import { isActive } from "../../utils/entity";
